add some tracking
